convert ReverseRequest members to simple types where possible

+++ src/test/java/de/komoot/photon/query/ReverseRequestFactoryTest.java
@@ -182,7 +182,7 @@ class ReverseRequestFactoryTest {
         Mockito.when(mockRequest.queryParams("limit")).thenReturn("51");
         ReverseRequestFactory reverseRequestFactory = new ReverseRequestFactory(Collections.singletonList("en"), "en");
         reverseRequest = reverseRequestFactory.create(mockRequest);
-        assertEquals(reverseRequest.getLimit().longValue(), 50);
+        assertEquals(reverseRequest.getLimit(), 50);
         Mockito.verify(mockRequest, Mockito.times(1)).queryParams("limit");
     }
